@@ -1,25 +1,17 @@
 package com.efe.gamedev.catacombs;
 
 import com.badlogic.gdx.Gdx;
+import com.badlogic.gdx.Screen;
 import com.badlogic.gdx.ScreenAdapter;
-import com.badlogic.gdx.graphics.Color;
+import com.badlogic.gdx.audio.Music;
+import com.badlogic.gdx.audio.Sound;
 import com.badlogic.gdx.graphics.GL20;
 import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.BitmapFont;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 import com.badlogic.gdx.graphics.glutils.ShapeRenderer;
-import com.badlogic.gdx.input.GestureDetector;
-import com.badlogic.gdx.math.MathUtils;
-import com.badlogic.gdx.math.Vector2;
-import com.badlogic.gdx.utils.Array;
-import com.badlogic.gdx.utils.TimeUtils;
-import com.badlogic.gdx.utils.viewport.ExtendViewport;
-import com.badlogic.gdx.utils.viewport.FillViewport;
-import com.badlogic.gdx.utils.viewport.FitViewport;
 import com.badlogic.gdx.utils.viewport.ScreenViewport;
 import com.badlogic.gdx.utils.viewport.Viewport;
-import com.efe.gamedev.catacombs.entities.Player;
-import com.efe.gamedev.catacombs.entities.SpeechBubble;
 import com.efe.gamedev.catacombs.entities.Word;
 import com.efe.gamedev.catacombs.util.ChaseCam;
 import com.efe.gamedev.catacombs.util.Constants;
@@ -32,12 +24,14 @@ import com.efe.gamedev.catacombs.util.Constants;
 public class GameplayScreen extends ScreenAdapter {
 
     private static final String TAG = GameplayScreen.class.getName();
-
+    //Game
+    public CatacombsGame game;
     //HUD
     private Viewport hudViewport;
     //Level
     private Levels levels;
     private Level level;
+    private int levelNumber = 0;
     //ShapeRenderer
     private ShapeRenderer renderer;
     //text setup
@@ -48,13 +42,48 @@ public class GameplayScreen extends ScreenAdapter {
 
     private ChaseCam chaseCam;
 
+    //add audio
+    //background music
+    private Music music_black_vortex;
+    private boolean play_black_vortex;
+    private Music music_enter_the_maze;
+    private boolean play_enter_the_maze;
+    private Music music_cave;
+    private boolean play_cave;
+
+    //default button sound
+    public Sound sound1;
+    //play buttons sound
+    public Sound sound2;
+    //get item sound
+    public Sound sound3;
+    //item pop-up sound
+    public Sound sound4;
+    //select speech bubble option
+    public Sound sound5;
+    //unlock and lock doors sound
+    public Sound sound6;
+    //enemy and play lasers sound
+    public Sound sound7;
+    //big thud sound for stalactites and spike pillars
+    public Sound sound8;
+    //explosion sound
+    public Sound sound9;
+
+    //set current level number
+    public GameplayScreen(CatacombsGame game, int levelNum) {
+        this.game = game;
+        levelNumber = levelNum;
+    }
+
     @Override
     public void show () {
         //initialize hud
         hudViewport = new ScreenViewport();
 
         levels = new Levels();
-        level = new Level(hudViewport, levels);
+        level = new Level(hudViewport, levels, this);
+        levels.currentLevel = levelNumber;
         levels.configureLevel(level);
 
         //initialize renderer
@@ -73,9 +102,32 @@ public class GameplayScreen extends ScreenAdapter {
         font2.getRegion().getTexture().setFilter(Texture.TextureFilter.Linear, Texture.TextureFilter.Linear);
         font3.getRegion().getTexture().setFilter(Texture.TextureFilter.Linear, Texture.TextureFilter.Linear);
 
-        //allow player to use external touch inputs
+        //allow level to use external touch inputs
         Gdx.input.setInputProcessor(level);
+        //initialize camera which follows player around the level
         chaseCam = new ChaseCam(level.viewport.getCamera(), level.getPlayer(), level.shake);
+        //initialize sound files
+        sound1 = Gdx.audio.newSound(Gdx.files.internal("sounds/nff_menu_05_a.wav"));
+        sound2 = Gdx.audio.newSound(Gdx.files.internal("sounds/nff_menu_a.wav"));
+        sound3 = Gdx.audio.newSound(Gdx.files.internal("sounds/nff_coin.wav"));
+        sound4 = Gdx.audio.newSound(Gdx.files.internal("sounds/nff_prompt.wav"));
+        sound5 = Gdx.audio.newSound(Gdx.files.internal("sounds/nff_confirm_02.wav"));
+        sound6 = Gdx.audio.newSound(Gdx.files.internal("sounds/nff_click_switch.wav"));
+        sound7 = Gdx.audio.newSound(Gdx.files.internal("sounds/nff_laser_gun.wav"));
+        sound8 = Gdx.audio.newSound(Gdx.files.internal("sounds/nff_thud.wav"));
+        sound9 = Gdx.audio.newSound(Gdx.files.internal("sounds/nff_explode.wav"));
+        //initialize background music files
+        music_black_vortex = Gdx.audio.newMusic(Gdx.files.internal("music/black_vortex.mp3"));
+        play_black_vortex = Constants.CURRENT_SONG.equals("black_vortex");
+        music_enter_the_maze = Gdx.audio.newMusic(Gdx.files.internal("music/enter_the_maze.mp3"));
+        play_enter_the_maze = Constants.CURRENT_SONG.equals("enter_the_maze");
+        music_cave = Gdx.audio.newMusic(Gdx.files.internal("music/chee_zee_cave.mp3"));
+        play_cave = Constants.CURRENT_SONG.equals("cave");
+
+        //play music
+        music_black_vortex.setLooping(false);
+        music_enter_the_maze.setLooping(false);
+        music_cave.setLooping(false);
     }
 
     @Override
@@ -100,14 +152,87 @@ public class GameplayScreen extends ScreenAdapter {
         font.dispose();
         font2.dispose();
         font3.dispose();
+        //dispose of sounds when app is closed
+        sound1.dispose();
+        sound2.dispose();
+        sound3.dispose();
+        sound4.dispose();
+        sound5.dispose();
+        sound6.dispose();
+        sound7.dispose();
+        sound8.dispose();
+        sound9.dispose();
+        //dispose of background music
+        music_black_vortex.dispose();
+        music_enter_the_maze.dispose();
+        music_cave.dispose();
     }
 
     @Override
     public void render (float delta) {
-        levels.update(level);
+        //first song
+        if (play_black_vortex && !music_black_vortex.isPlaying()) {
+            music_black_vortex.play();
+            play_black_vortex = false;
+        }
+        music_black_vortex.setOnCompletionListener(new Music.OnCompletionListener() {
+            @Override
+            public void onCompletion(Music music) {
+                play_enter_the_maze = true;
+                Constants.CURRENT_SONG = "enter_the_maze";
+            }
+        }
+        );
+        //second song
+        if (play_enter_the_maze && !music_enter_the_maze.isPlaying()) {
+            music_enter_the_maze.play();
+            play_enter_the_maze = false;
+        }
+        music_enter_the_maze.setOnCompletionListener(new Music.OnCompletionListener() {
+            @Override
+            public void onCompletion(Music music) {
+                play_cave = true;
+                Constants.CURRENT_SONG = "cave";
+            }
+        }
+        );
+        //third song
+        if (play_cave && !music_cave.isPlaying()) {
+            music_cave.play();
+            play_cave = false;
+        }
+        music_cave.setOnCompletionListener(new Music.OnCompletionListener() {
+            @Override
+            public void onCompletion(Music music) {
+                play_black_vortex = true;
+                Constants.CURRENT_SONG = "black_vortex";
+            }
+        }
+        );
+        //make songs quieter when paused
+        if (level.inventory.paused || level.defeat || level.victory) {
+            music_black_vortex.setVolume(0.5f);
+            music_enter_the_maze.setVolume(0.5f);
+            music_cave.setVolume(0.5f);
+        } else {
+            music_black_vortex.setVolume(1f);
+            music_enter_the_maze.setVolume(1f);
+            music_cave.setVolume(1f);
+        }
+
+        levels.update(level, delta);
         level.update(delta);
+        //update camera
         chaseCam.update();
         chaseCam.shake = level.shake;
+        //change camera shake intensity
+        if (levels.currentLevel == 14 && level.currentBubble > 34) {
+            chaseCam.shakeDilation = 3f;
+        } else {
+            chaseCam.shakeDilation = 1f;
+        }
+        level.cameraPosition.x = (chaseCam.camera.position.x);
+        level.cameraPosition.y = (chaseCam.camera.position.y);
 
         Gdx.gl.glClearColor(Constants.BACKGROUND_COLOR.r, Constants.BACKGROUND_COLOR.g, Constants.BACKGROUND_COLOR.b, 1);
         Gdx.gl.glClear(GL20.GL_COLOR_BUFFER_BIT);
@@ -127,7 +252,7 @@ public class GameplayScreen extends ScreenAdapter {
         batch.begin();
 
         level.renderText(batch, font, font2);
-        if (!level.inventory.paused) {
+        if (!level.inventory.paused && !level.victory && !level.defeat) {
             for (Word word: level.words) {
                 word.renderCollectedText(batch, font, hudViewport);
             }
@@ -136,9 +261,22 @@ public class GameplayScreen extends ScreenAdapter {
         batch.end();
     }
 
+    public void showMenuScreen(int currentLevel) {
+        //Constants.SONG_POSITION = music_black_vortex.isPlaying() ? music_black_vortex.getPosition() : (music_enter_the_maze.isPlaying() ? music_enter_the_maze.getPosition() : (music_cave.isPlaying() ? music_cave.getPosition() : 0));
+        //dispose of background music
+        music_black_vortex.dispose();
+        music_enter_the_maze.dispose();
+        music_cave.dispose();
+        //set screen
+        game.showMenuScreen(currentLevel);
+    }
+
     @Override
     public void pause() {
-
+        //pause level
+        if (!level.victory && !level.defeat) {
+            level.inventory.paused = true;
+        }
     }
 
     @Override
