@@ -1,13 +1,13 @@
 package com.efe.gamedev.catacombs;
 
 import com.badlogic.gdx.Gdx;
+import com.badlogic.gdx.Input;
 import com.badlogic.gdx.InputAdapter;
 import com.badlogic.gdx.graphics.Color;
 import com.badlogic.gdx.graphics.GL20;
 import com.badlogic.gdx.graphics.g2d.BitmapFont;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 import com.badlogic.gdx.graphics.glutils.ShapeRenderer;
-import com.badlogic.gdx.input.GestureDetector;
 import com.badlogic.gdx.math.MathUtils;
 import com.badlogic.gdx.math.Vector2;
 import com.badlogic.gdx.utils.Array;
@@ -15,13 +15,19 @@ import com.badlogic.gdx.utils.DelayedRemovalArray;
 import com.badlogic.gdx.utils.TimeUtils;
 import com.badlogic.gdx.utils.viewport.ExtendViewport;
 import com.badlogic.gdx.utils.viewport.Viewport;
+import com.efe.gamedev.catacombs.entities.Boss;
 import com.efe.gamedev.catacombs.entities.Catacomb;
 import com.efe.gamedev.catacombs.entities.Chest;
+import com.efe.gamedev.catacombs.entities.Door;
+import com.efe.gamedev.catacombs.entities.Electricity;
 import com.efe.gamedev.catacombs.entities.Exit;
 import com.efe.gamedev.catacombs.entities.Guard;
 import com.efe.gamedev.catacombs.entities.Inventory;
 import com.efe.gamedev.catacombs.entities.Item;
+import com.efe.gamedev.catacombs.entities.Lever;
+import com.efe.gamedev.catacombs.entities.Prisoner;
 import com.efe.gamedev.catacombs.entities.Puzzle;
+import com.efe.gamedev.catacombs.entities.Sign;
 import com.efe.gamedev.catacombs.entities.Target;
 import com.efe.gamedev.catacombs.entities.Word;
 import com.efe.gamedev.catacombs.entities.levelVerdict;
@@ -41,9 +47,10 @@ public class Level extends InputAdapter {
 
     private static final String TAG = Level.class.getName();
 
+    //gameplay screen
+    public GameplayScreen gameplayScreen;
     //all levels
     public Levels superior;
-
     //viewport
     public Viewport viewport;
     //hudViewport
@@ -53,6 +60,8 @@ public class Level extends InputAdapter {
     //seperate touch position for buttons and things
     public Vector2 touchPosition;
     public Vector2 lookPosition;
+    public boolean pressDown;
+    public boolean pressUp;
     //allows player to touch stuff or not
     public boolean touchLocked;
 
@@ -68,6 +77,9 @@ public class Level extends InputAdapter {
     //speechBubbles
     public Array<SpeechBubble> speechBubbles;
     public int currentBubble;
+    //save last bubble seen
+    public int lastSeenBubble;
+    //variables controlling sight of speechBubbles
     public boolean continueBubbles = true;
     public boolean show = true;
     //Catacomb Chambers
@@ -75,7 +87,8 @@ public class Level extends InputAdapter {
     public int currentCatacomb;
     //items
     public DelayedRemovalArray<Item> items;
-    public boolean[] collectedItems = new boolean[]{false, false, false, false, false, false};
+    //A boolean array containing data that stores which items have been collected. True means that the player has collected that item before and false means that the player has never seen the item before.
+    public boolean[] collectedItems = new boolean[]{false, false, false, false, false, false, false, false, false, false, false, false, false, false, false, false, false, false};
     public Array<String> collectedItemTypes;
     //player's Inventory
     public Inventory inventory;
@@ -89,6 +102,19 @@ public class Level extends InputAdapter {
     public Target targetBox;
     //guards
     public Array<Guard> guards;
+    //Levers
+    public Array<Lever> levers;
+    //Doors
+    public Array<Door> doors;
+    public boolean usedDoor;
+    //Signs
+    public Array<Sign> signs;
+    //Electrical wire puzzles
+    public Array<Electricity> wires;
+    //prisoners
+    public Array<Prisoner> prisoners;
+    //bosses
+    public Array<Boss> bosses;
 
     //torch light
     public long startTime;
@@ -96,25 +122,44 @@ public class Level extends InputAdapter {
     private static final float TORCH_MOVEMENT_DISTANCE = 5;
     private static final float TORCH_PERIOD = 2f;
     public boolean levelStarted = false;
+    public boolean torchFade;
+    public boolean torchUp;
     //alarm
     public boolean alarm;
     private float alarmLight;
     public boolean shake;
+    //camera move
+    public Vector2 cameraPosition;
 
-    public Level (Viewport hudViewport, Levels superior) {
+    public Level (Viewport hudViewport, Levels superior, GameplayScreen gameplayScreen) {
         this.superior = superior;
+        this.gameplayScreen = gameplayScreen;
         //viewport parameters
         viewport = new ExtendViewport(Constants.WORLD_SIZE, Constants.WORLD_SIZE);
         viewportPosition = new Vector2();
         touchPosition = new Vector2();
         lookPosition = new Vector2();
+        pressDown = false;
+        pressUp = false;
         touchLocked = true;
         //initiate some stuff
         this.hudViewport = hudViewport;
         alarm = false;
         shake = false;
+        cameraPosition = new Vector2();
+        //all items
         collectedItemTypes = new Array<String>();
-        collectedItemTypes.add("key"); collectedItemTypes.add("dagger"); collectedItemTypes.add("gold"); collectedItemTypes.add("phone"); collectedItemTypes.add("diamond"); collectedItemTypes.add("stungun");
+        collectedItemTypes.add("key"); collectedItemTypes.add("dagger"); collectedItemTypes.add("gold");
+        collectedItemTypes.add("phone"); collectedItemTypes.add("diamond"); collectedItemTypes.add("stungun");
+        collectedItemTypes.add("sapphire"); collectedItemTypes.add("ruby"); collectedItemTypes.add("emerald");
+        collectedItemTypes.add("invisibility"); collectedItemTypes.add("ghost"); collectedItemTypes.add("doubleKey");
+        collectedItemTypes.add("shield"); collectedItemTypes.add("bomb"); collectedItemTypes.add("fire");
+        collectedItemTypes.add("shock"); collectedItemTypes.add("disguise"); collectedItemTypes.add("spear");
+        if (gameplayScreen.game.getFurthestLevel() > superior.currentLevel) {
+            lastSeenBubble = 60;
+        } else {
+            lastSeenBubble = 0;
+        }
 
         victory = false;
         defeat = false;
@@ -126,6 +171,7 @@ public class Level extends InputAdapter {
         targetBox = new Target();
         //get player
         player = new Player(new Vector2(100, 150), viewportPosition, this);
+        Gdx.input.setCatchBackKey(true);
     }
 
     @Override
@@ -135,6 +181,8 @@ public class Level extends InputAdapter {
             viewportPosition = viewport.unproject(new Vector2(screenX, screenY));
         }
         touchPosition = viewport.unproject(new Vector2(screenX, screenY));
+        pressDown = true;
+        pressUp = false;
         //give objects and player the touch position when player touches screen
         player.viewportPosition = viewportPosition;
         for (Item item : items) {
@@ -142,12 +190,17 @@ public class Level extends InputAdapter {
         }
 
         //change speechBubble
+        //make sure you are not touching on the pause button
         if (touchPosition.dst(new Vector2((player.getPosition().x - viewport.getWorldWidth() / 2) + 2, player.getPosition().y - viewport.getWorldHeight() / 2)) > 18 && !speechBubbles.get(currentBubble).textLoaded) {
             speechBubbles.get(currentBubble).typeSpeed = 2;
         }
-        if (speechBubbles.get(currentBubble).Options == 0 && continueBubbles && currentBubble < speechBubbles.size - 1 && speechBubbles.get(currentBubble).textLoaded && !inventory.paused && touchPosition.dst(new Vector2((player.getPosition().x - viewport.getWorldWidth() / 2) + 2, player.getPosition().y - viewport.getWorldHeight() / 2)) > 18) {
-            speechBubbles.get(currentBubble).typeSpeed = 4;
+        if (speechBubbles.get(currentBubble).Options == 0 && continueBubbles && currentBubble < speechBubbles.size - 1 && speechBubbles.get(currentBubble).textLoaded && !inventory.paused && touchPosition.dst(new Vector2((player.getPosition().x - viewport.getWorldWidth() / 2) + 2, player.getPosition().y - viewport.getWorldHeight() / 2)) > 18 && (speechBubbles.get(currentBubble).speechBubbleSkip == -1 || touchPosition.dst(new Vector2(speechBubbles.get(currentBubble).position.x + speechBubbles.get(currentBubble).width + 59.5f + speechBubbles.get(currentBubble).targetOffset, (speechBubbles.get(currentBubble).height == 20 ? ((speechBubbles.get(currentBubble).position.y + 7.5f) + 10) : (speechBubbles.get(currentBubble).position.y + 7.5f)))) > 30)) {
+            speechBubbles.get(currentBubble).typeSpeed = 3;
             currentBubble += 1;
+            //only update lastSeenBubble if currentBubble is more or equal to it.
+            if (currentBubble >= lastSeenBubble) {
+                lastSeenBubble = currentBubble;
+            }
         }
         //check to see if you want the player to jump when you tap on the screen.
         if (!touchLocked && !inventory.paused) {
@@ -163,26 +216,88 @@ public class Level extends InputAdapter {
                 guards.get(i).guardEnergy -= 1;
             }
         }
-
+        //duck lasers
         if (superior.currentLevel == 5 && currentBubble == 35) {
             if (touchPosition.x > this.getPlayer().getPosition().x - Constants.HEAD_SIZE && touchPosition.x < (this.getPlayer().getPosition().x - Constants.HEAD_SIZE) + Constants.PLAYER_WIDTH * 2f && touchPosition.y > this.getPlayer().getPosition().y + Constants.HEAD_SIZE - Constants.PLAYER_HEIGHT * 2.5f && touchPosition.y < this.getPlayer().getPosition().y + Constants.HEAD_SIZE && player.energy > 0 && !player.recoverE) {
                 player.duck = true;
             }
         }
+        //shoot lasers back at guard
+        if (superior.currentLevel == 8 && currentBubble == 9 && player.heldItem.itemType.equals("stungun") && player.energy >= 20 && !defeat && speechBubbles.get(currentBubble).textLoaded) {
+            player.heldItem.stungun.shoot = true;
+            player.energy = 0;
+        }
+        //shoot lasers at boss during boss battle
+        if (touchPosition.dst(new Vector2(player.getPosition().x - Constants.HEAD_SIZE * 1.9f, player.getPosition().y - viewport.getWorldHeight() / 2f)) < 30 && superior.currentLevel == 14 && currentBubble > 50 && player.heldItem.itemType.equals("stungun") && player.energy >= 20 && !defeat && speechBubbles.get(currentBubble).textLoaded) {
+            player.heldItem.stungun.shoot = true;
+            player.energy = 0;
+        }
+
         //returns the position at which your finger touches device screen
         return super.touchDown(screenX, screenY, pointer, button);
+    }
+
+    @Override
+    public boolean touchDragged(int screenX, int screenY, int pointer) {
+        //make player duck in boss fight
+        if (superior.currentLevel == 14 && currentBubble > 17 && currentBubble < 20) {
+            player.duck = (viewport.unproject(new Vector2(screenX, screenY)).y < touchPosition.y - 10) && (player.position.y < catacombs.get(currentCatacomb).position.y + catacombs.get(currentCatacomb).wallThickness + 64);
+            if ((player.position.y < catacombs.get(currentCatacomb).position.y + catacombs.get(currentCatacomb).wallThickness + 64) && !player.duck && ((viewport.unproject(new Vector2(screenX, screenY)).y > touchPosition.y + 10))) {
+                player.setVelocity(new Vector2(0, 200f));
+            }
+        }
+        //check if player is dragging bomb
+        if (inventory.touchItem("bomb") != -1 && !inventory.newItem) {
+            inventory.dragItem = true;
+        } else {
+            inventory.dragItem = false;
+        }
+        //set bomb to player's touch position when bomb is being dragged
+        if (inventory.dragItem) {
+            inventory.selectedItem = inventory.touchItem("bomb");
+            inventory.inventoryItems.get(inventory.touchItem("bomb")).position = viewport.unproject(new Vector2(screenX, screenY));
+        }
+        //keeps returning the position at which your finger is touching the device screen when you are dragging your finger
+        return super.touchDragged(screenX, screenY, pointer);
     }
 
     @Override
     public boolean touchUp (int screenX, int screenY, int pointer, int button) {
 
         lookPosition = viewport.unproject(new Vector2(screenX, screenY));
+        //boss fight stop dragging
+        if (superior.currentLevel == 14) {
+            player.duck = false;
+        }
+        //stop dragging
+        if (inventory.dragItem) {
+            //check if item is out of inventory and if item is in currentCatacomb's bounds
+            Catacomb catacomb = catacombs.get(currentCatacomb);
+            if ((inventory.inventoryItems.get(inventory.touchItem("bomb")).position.x < inventory.position.x || inventory.inventoryItems.get(inventory.touchItem("bomb")).position.x > inventory.position.x + inventory.width || inventory.inventoryItems.get(inventory.touchItem("bomb")).position.y > inventory.position.y + inventory.height) && inventory.inventoryItems.get(inventory.touchItem("bomb")).position.x > catacomb.position.x + 50 && inventory.inventoryItems.get(inventory.touchItem("bomb")).position.x < (catacomb.position.x + catacomb.width) - 50 && inventory.inventoryItems.get(inventory.touchItem("bomb")).position.y > catacomb.position.y + 35 && inventory.inventoryItems.get(inventory.touchItem("bomb")).position.y < catacomb.position.y + catacomb.height - 35) {
+                items.add(new Item(new Vector2(lookPosition), viewportPosition, "bomb"));
+                items.get(items.size - 1).bomb.triggered = true;
+                items.get(items.size - 1).bomb.currentCatacomb = currentCatacomb;
+                inventory.deleteCurrentItem();
+            }
+            inventory.dragItem = false;
+        }
+        //stop pressing
+        pressDown = false;
+        pressUp = true;
         //makes it so that in level 6, at a certain spot in the level, when you stop touching the player, it stops ducking lasers.
         if (superior.currentLevel == 5 && currentBubble == 35) {
             player.duck = false;
         }
         //returns the position at which your finger stops touching device screen
         return super.touchUp(screenX, screenY, pointer, button);
+    }
+
+    @Override
+    public boolean keyDown(int keycode) {
+        if (keycode == Input.Keys.BACK) {
+            inventory.paused = true;
+        }
+        return false;
     }
 
     public void update (float delta) {
@@ -194,6 +309,15 @@ public class Level extends InputAdapter {
                 player.update(delta);
                 for (Guard guard : guards) {
                     guard.update(delta);
+                }
+                for (Prisoner prisoner : prisoners) {
+                    prisoner.update(delta);
+                }
+                for (Boss boss : bosses) {
+                    boss.update(delta);
+                }
+                for (Lever lever : levers) {
+                    lever.update();
                 }
                 exitDoor.update(delta);
                 if (targetBox.target) {
@@ -219,11 +343,13 @@ public class Level extends InputAdapter {
                     item.shadowOffset.set(new Vector2((player.getPosition().x - item.position.x) / -10, -2));
                     //if collected, let score or inventory pick it up
                     if (item.collected) {
+                        gameplayScreen.sound3.play();
                         //if diamond, add to ScoreDiamonds (records how many diamonds you have collected in the level), else, add to inventory items
                         if (item.itemType.equals("diamond")) {
                             inventory.scoreDiamonds.add(new Diamond(new Vector2()));
                         } else {
                             inventory.inventoryItems.add(new Item(new Vector2(item.position.x, item.position.y), viewportPosition, item.itemType));
+                            inventory.inventoryItems.get(inventory.inventoryItems.size - 1).potion.full = item.potion.full;
                         }
                         items.removeIndex(i);
                     }
@@ -234,12 +360,33 @@ public class Level extends InputAdapter {
             }
             inventory.update();
             //update torch light
-            if (torchLight < 60f && !levelStarted) {
-                torchLight += 0.5f;
-                startTime = TimeUtils.nanoTime();
+            if (!torchFade) {
+                if (torchLight < 60f && !levelStarted) {
+                    torchLight += 0.5f;
+                    startTime = TimeUtils.nanoTime();
+                } else {
+                    levelStarted = true;
+                    updateTorch();
+                }
             } else {
-                levelStarted = true;
-                updateTorch();
+                //used for opening doors
+                //fade torchlight
+                if (torchLight > 0f && !torchUp) {
+                    torchLight -= 1f;
+                }
+                //when torchlight is very low, make it become larger again and move player to the location of the exitDoorIndex.
+                if (torchLight <= 0f && !torchUp) {
+                    torchUp = true;
+                }
+                //move torchlight up again
+                if (torchLight <= 60f && torchUp) {
+                    torchLight += 1f;
+                }
+                //reset torchlight
+                if (torchLight > 59 && torchUp) {
+                    torchFade = false;
+                    torchUp = false;
+                }
             }
         }
         //update alarm light
@@ -294,8 +441,28 @@ public class Level extends InputAdapter {
             puzzle.render(renderer);
             puzzle.renderHint(renderer);
         }
+        //Doors
+        for (Door door : doors) {
+            door.render(renderer);
+        }
+        //Signs
+        for (Sign sign : signs) {
+            sign.render(renderer);
+        }
+        //Electricity
+        for (Electricity wire : wires) {
+            wire.render(renderer, this);
+        }
+        //prisoners
+        for (Prisoner prisoner : prisoners) {
+            prisoner.render(renderer);
+        }
         //player
         player.render(renderer);
+        //boss
+        for (Boss boss : bosses) {
+            boss.render(renderer);
+        }
         //guards
         for (Guard guard : guards) {
             guard.render(renderer);
@@ -307,13 +474,13 @@ public class Level extends InputAdapter {
         for (Catacomb catacomb : catacombs) {
             catacomb.renderButton(renderer);
         }
-        //items
-        for (Item item : items) {
-            item.render(renderer, this);
-        }
         //chests
         for (Chest chest : chests) {
             chest.render(renderer);
+        }
+        //Levers
+        for (Lever lever : levers) {
+            lever.render(renderer);
         }
         //render exit
         if (exitDoor.unlocked) {
@@ -321,6 +488,14 @@ public class Level extends InputAdapter {
         }
         //render exit button
         exitDoor.renderButton(renderer);
+        //items
+        for (Item item : items) {
+            item.render(renderer, this);
+        }
+        //red covering over catacombs
+        for (Catacomb catacomb : catacombs) {
+            catacomb.renderRedCatacomb(renderer);
+        }
         renderer.end();
         Gdx.gl.glDisable(GL20.GL_BLEND);
     }
@@ -337,8 +512,9 @@ public class Level extends InputAdapter {
         renderer.set(ShapeRenderer.ShapeType.Filled);
         for (int i = 0; i < viewport.getWorldWidth() / 4f; i+=1) {
             renderer.setColor(new Color(Color.BLACK.r, Color.BLACK.g, Color.BLACK.b, i / (torchLight)));
-            rectBorder(renderer, new Vector2(player.getPosition().x - (i * 2) - Constants.HEAD_SIZE, player.getPosition().y - (i * 2) - Constants.HEAD_SIZE), 20 + (i * 4), 10 + (i * 4));
+            rectBorder(renderer, new Vector2(cameraPosition.x - (i * 2) - Constants.HEAD_SIZE, cameraPosition.y - (i * 2) - Constants.HEAD_SIZE), 20 + (i * 4), 10 + (i * 4));
         }
+
         //alarm
         if (alarm) {
             for (int i = 47; i < viewport.getWorldWidth() / 4f; i+=1) {
@@ -346,8 +522,13 @@ public class Level extends InputAdapter {
                 rectBorder(renderer, new Vector2((player.getPosition().x - (i * 2) - Constants.HEAD_SIZE - 2) - ((viewport.getWorldWidth() - 213) * 0.5f), (player.getPosition().y - (i * 2) - Constants.HEAD_SIZE) + 28), (viewport.getWorldWidth() - 193) + (i * 4), -40 + (i * 4));
             }
         }
+        //target boxes
         if (targetBox.target && !inventory.paused) {
             targetBox.render(renderer);
+        }
+        //boss Spike Pillar warnings
+        for (Boss boss : bosses) {
+            boss.renderSpikeWarning(renderer);
         }
         if (!victory && !defeat) {
             //Inventory for player
@@ -380,19 +561,18 @@ public class Level extends InputAdapter {
             if (show && !inventory.paused && !inventory.newItem) {
                 speechBubbles.get(currentBubble).renderText(batch, font, hudViewport);
             }
-            //TODO: Make it so the player can make a custom name for the character
             if (speechBubbles.get(currentBubble).target.equals("Player:")) {
                 if (speechBubbles.get(currentBubble).textLoaded) {
                     player.mouthState = Enums.MouthState.NORMAL;
                 } else {
-                    if (inventory.speechButton.speech && show) {
+                    if (inventory.speechButton.speech && show && !inventory.newItem) {
                         player.mouthState = Enums.MouthState.TALKING;
                     } else {
                         player.mouthState = Enums.MouthState.NORMAL;
                     }
                 }
             }
-            //guards talking states
+            //guards' talking states
             for (int i = 0; i < guards.size; i++) {
                 if (show && speechBubbles.get(currentBubble).target.equals("Guard " + (i+1) + ":")) {
                     if (speechBubbles.get(currentBubble).textLoaded) {
@@ -402,8 +582,24 @@ public class Level extends InputAdapter {
                     }
                 }
             }
+            //prisoners' talking states
+            if (show && (speechBubbles.get(currentBubble).target.equals("Prisoner 48H:") || speechBubbles.get(currentBubble).target.equals("Prisoner 12E:"))) {
+                if (speechBubbles.get(currentBubble).textLoaded) {
+                    prisoners.get(0).mouthState = Enums.MouthState.NORMAL;
+                } else {
+                    prisoners.get(0).mouthState = Enums.MouthState.TALKING;
+                }
+            }
+            //boss's talking states
+            if (show && (speechBubbles.get(currentBubble).target.equals("Boss:"))) {
+                if (speechBubbles.get(currentBubble).textLoaded) {
+                    bosses.get(0).mouthState = Enums.MouthState.NORMAL;
+                } else {
+                    bosses.get(0).mouthState = Enums.MouthState.TALKING;
+                }
+            }
         }
-        if (inventory.paused || inventory.newItem) {
+        if (inventory.paused || inventory.newItem || !inventory.message.equals("")) {
             inventory.renderText(batch, font2, font, hudViewport);
         }
         if (victory || defeat) {
